[untested] Ok I think I fixed the dupe bug
So half of this was wrapped in a scheduler - perhaps to re-add the fake
potato item after a crate was "crafted." Problem is, he was overwriting
the results when calling updateInventory... and perhaps some other
stuff.

I'll see if this fixes it first, then I can go about a better way to
re-add the potato.